fixed float within range for py3
Fixed games so they pass the tests for both py2 and py3. Most
of the changes were due to using floats for range() which isn't
permitted in py3.

+++ ple/games/catcher.py
@@ -101,7 +101,7 @@ class Fruit(pygame.sprite.Sprite):
         y = self.rng.choice(
             range(
                 self.size,
-                self.SCREEN_HEIGHT / 2,
+                int(self.SCREEN_HEIGHT / 2),
                 self.size))
 
         self.rect.center = (x, -1 * y)
